refactor(error-tracking): move notification types to core
