+++ rust/cymbal/src/core/mod.rs
@@ -10,5 +10,6 @@ pub mod ids;
 pub mod metric_consts;
 pub mod resolver;
 pub mod sanitize;
+pub mod shutdown;
 pub mod symbolication;
 pub mod types;
